requirements.txt
# === Flask stack ===
flask==2.0.3
flask-jwt-extended==4.3.1
flask-swagger-ui==3.36.0
flask-cors==3.0.10
Werkzeug==2.0.3
pyjwt==2.9.0
python-dotenv==1.0.1

# === CV / OCR / Mongo ===
yolov5==7.0.14
easyocr==1.7.2
pymupdf==1.25.3
pymongo==4.3.3
pytesseract==0.3.10

# === Hugging Face Transformers for NER ===
transformers==4.34.1
datasets==2.14.6
seqeval==1.2.2
torch>=1.13.0  # Pour les modèles HF, requis pour transformers
accelerate==0.24.1  # Optimisation du training HF
scikit-learn==1.3.2  # Utilitaire métriques si besoin

# === Tokenizer ===
sentencepiece==0.1.99
protobuf==3.20.3  # Compatible avec certaines versions HF

# === Utility ===
deepmerge==1.1.1